Code now working. Removed some checks and will now run speedtest.
