+++ gen_reads.py
@@ -24,7 +24,6 @@ import pickle
 import numpy as np
 import argparse
 import pathlib
-import pdb
 
 from source.input_checking import check_file_open, is_in_range
 from source.ref_func import index_ref, read_ref
@@ -94,7 +93,7 @@ def main(raw_args=None):
                              'things like read locations, variant positions, error positions, etc, '
                              'should all be the same.')
     # TODO check if this argument does anything at all. Near as I can tell the results are ALWAYS gzipped.
-    parser.add_argument('--gz', required=False, action='store_true', default=False, help='gzip output FQ and VCF')
+    # parser.add_argument('--gz', required=False, action='store_true', default=False, help='gzip output FQ and VCF')
     parser.add_argument('--no-fastq', required=False, action='store_true', default=False,
                         help='bypass fastq generation')
     parser.add_argument('--discard-offtarget', required=False, action='store_true', default=False,
